feat(ui): move lists from settings to sidebar (#1263)
