+++ desk/src/components/desk/settings/teams/TeamInfo.vue
@@ -3,7 +3,7 @@
 		<div class="flow-root h-[72px] shrink-0 py-[22px] px-[16px]">
 			<div class="float-left">
 				<router-link
-					:to="`/settings/teams`"
+					:to="`/teams`"
 					class="my-1 flex select-none flex-row items-center space-x-1 stroke-gray-600 text-sm text-gray-600 hover:stroke-gray-700 hover:text-gray-700"
 					role="button"
 				>
