+++ desk/src/components/desk/sidebar/SidebarLink.vue
@@ -1,43 +1,32 @@
 <template>
-  <router-link :to="{ name: to }">
+  <div
+    class="flex h-7 cursor-pointer items-center rounded-lg px-1.5 text-gray-700 transition-all duration-300 ease-in-out"
+    :class="{
+      'w-full': sidebarStore.isExpanded,
+      'w-7': !sidebarStore.isExpanded,
+      'bg-gray-200': isActive,
+      'hover:bg-gray-300': isActive,
+      'hover:bg-gray-100': !isActive,
+    }"
+    @click="handle"
+  >
+    <component :is="icon" class="h-4 w-4 shrink-0"></component>
     <div
-      class="flex h-7 cursor-pointer items-center rounded-lg px-1.5 text-gray-700 transition-all duration-300 ease-in-out"
+      class="ml-2 shrink-0 text-sm transition-all duration-300 ease-in-out"
       :class="{
-        'w-full': sidebarStore.isExpanded,
-        'w-7': !sidebarStore.isExpanded,
-        'bg-gray-200': isActive,
-        'text-gray-900': isActive,
-        'hover:bg-gray-300': isActive,
-        'hover:bg-gray-100': !isActive,
+        'opacity-100': sidebarStore.isExpanded,
+        'opacity-0': !sidebarStore.isExpanded,
+        '-z-50': !sidebarStore.isExpanded,
       }"
     >
-      <component
-        :is="icon"
-        v-show="!isActive"
-        class="h-4 w-4 shrink-0"
-      ></component>
-      <component
-        :is="iconActive"
-        v-show="isActive"
-        class="h-4 w-4 shrink-0"
-      ></component>
-      <div
-        class="ml-2 shrink-0 text-sm transition-all duration-300 ease-in-out"
-        :class="{
-          'opacity-100': sidebarStore.isExpanded,
-          'opacity-0': !sidebarStore.isExpanded,
-          '-z-50': !sidebarStore.isExpanded,
-        }"
-      >
-        {{ label }}
-      </div>
+      {{ label }}
     </div>
-  </router-link>
+  </div>
 </template>
 
 <script setup lang="ts">
-import { computed, toRefs } from "vue";
-import { useRoute } from "vue-router";
+import { computed } from "vue";
+import { useRoute, useRouter } from "vue-router";
 import { useSidebarStore } from "@/stores/sidebar";
 
 const props = defineProps({
@@ -45,23 +34,33 @@ const props = defineProps({
     type: String,
     required: true,
   },
-  to: {
-    type: String,
-    required: true,
-  },
   icon: {
     type: Object,
     required: true,
   },
-  iconActive: {
-    type: Object,
+  to: {
+    type: String,
     required: false,
-    default: null,
+    default: "",
+  },
+  handler: {
+    type: Function,
+    required: false,
+    default: () => true,
   },
 });
 
-const { to } = toRefs(props);
 const route = useRoute();
+const router = useRouter();
 const sidebarStore = useSidebarStore();
-const isActive = computed(() => to.value.includes(route.name.toString()));
+const isActive = computed(() => props.to.includes(route.name.toString()));
+
+function handle() {
+  props.handler();
+  if (props.to) {
+    router.push({
+      name: props.to,
+    });
+  }
+}
 </script>
