+++ desk/src/pages/desk/agent/AgentList.vue
@@ -1,0 +1,108 @@
+<template>
+  <div class="flex flex-col">
+    <PageTitle title="Agents">
+      <template #right>
+        <Button
+          icon-left="plus"
+          label="New agent"
+          class="bg-gray-900 text-white hover:bg-gray-800"
+          @click="isDialogVisible = !isDialogVisible"
+        />
+      </template>
+    </PageTitle>
+    <HelpdeskTable
+      class="grow"
+      :columns="columns"
+      :data="contacts.list?.data || []"
+      row-key="name"
+      :hide-checkbox="true"
+      :hide-column-selector="true"
+    >
+      <template #name="{ data }">
+        <div class="flex items-center justify-between">
+          <div class="flex items-center gap-2">
+            <Avatar
+              :label="data.name"
+              :image-u-r-l="data.user_image"
+              size="sm"
+            />
+            <div class="line-clamp-1">{{ data.full_name }}</div>
+          </div>
+          <Badge v-if="!data.is_active">Inactive</Badge>
+        </div>
+      </template>
+      <template #row-extra="{ data }">
+        <div class="cursor-pointer text-xs" @click="toTickets(data.name)">
+          Tickets &rightarrow;
+        </div>
+      </template>
+    </HelpdeskTable>
+    <ListNavigation class="p-3" v-bind="contacts" />
+    <AddNewAgentsDialog
+      :show="isDialogVisible"
+      @close="isDialogVisible = false"
+    />
+  </div>
+</template>
+<script setup lang="ts">
+import { ref } from "vue";
+import { useRouter } from "vue-router";
+import { Avatar, Badge } from "frappe-ui";
+import { AGENT_PORTAL_TICKET_LIST } from "@/router";
+import { createListManager } from "@/composables/listManager";
+import { useListFilters } from "@/composables/listFilters";
+import AddNewAgentsDialog from "@/components/desk/global/AddNewAgentsDialog.vue";
+import PageTitle from "@/components/PageTitle.vue";
+import HelpdeskTable from "@/components/HelpdeskTable.vue";
+import ListNavigation from "@/components/ListNavigation.vue";
+
+const router = useRouter();
+const filters = useListFilters();
+const isDialogVisible = ref(false);
+const columns = [
+  {
+    title: "Name",
+    colKey: "name",
+    colClass: "w-1/3",
+  },
+  {
+    title: "Email",
+    colKey: "email",
+    colClass: "w-1/3",
+  },
+  {
+    title: "Username",
+    colKey: "username",
+  },
+];
+
+const contacts = createListManager({
+  doctype: "HD Agent",
+  fields: [
+    "name",
+    "is_active",
+    "user.full_name",
+    "user.user_image",
+    "user.email",
+    "user.username",
+  ],
+  auto: true,
+});
+
+function toTickets(user: string) {
+  const q = filters.toQuery([
+    {
+      fieldname: "_assign",
+      filter_type: "is",
+      value: user,
+    },
+  ]);
+
+  router.push({
+    name: AGENT_PORTAL_TICKET_LIST,
+    query: {
+      q,
+    },
+  });
+}
+</script>
