--- desk/src/pages/desk/settings/canned_response/CannedResponse.vue
@@ -1,24 +1,0 @@
-<template>
-	<div class="flex flex-col">
-		<CannedResponseInfo
-			class="shrink-0 border-r border-[#F4F5F6]"
-			:canned_response="canned_responseId"
-		/>
-	</div>
-</template>
-
-<script>
-import CannedResponseInfo from "@/components/desk/settings/canned_responses/CannedResponseInfo.vue"
-export default {
-	name: "CannedResponse",
-	props: ["canned_responseId"],
-	components: {
-		CannedResponseInfo,
-	},
-	setup() {},
-	mounted() {
-		this.$event.emit("set-selected-setting", "Canned Responses")
-		this.$event.emit("show-top-panel-actions-settings", "CannedResponse")
-	},
-}
-</script>
